@@ -11,25 +11,56 @@
 	<dependencyManagement>
 		<dependencies>
 			<dependency>
+				<groupId>io.zeebe</groupId>
+				<artifactId>zeebe-test-container</artifactId>
+				<version>3.3.0</version>
+			</dependency>
+			<dependency>
 				<groupId>io.camunda</groupId>
 				<artifactId>zeebe-bom</artifactId>
-				<version>1.2.3</version>
+				<version>1.2.9</version>
 				<scope>import</scope>
 				<type>pom</type>
+			</dependency>
+			<dependency>
+				<groupId>org.junit</groupId>
+				<artifactId>junit-bom</artifactId>
+				<version>5.8.2</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+			<dependency>
+				<groupId>org.testcontainers</groupId>
+				<artifactId>testcontainers-bom</artifactId>
+				<version>1.16.3</version>
+				<type>pom</type>
+				<scope>import</scope>
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
 
 	<dependencies>
-
 		<dependency>
-			<groupId>io.camunda</groupId>
-			<artifactId>zeebe-client-java</artifactId>
+			<groupId>io.zeebe</groupId>
+			<artifactId>zeebe-test-container</artifactId>
+			<scope>test</scope>
 		</dependency>
 
 		<dependency>
 			<groupId>io.camunda</groupId>
-			<artifactId>zeebe-test</artifactId>
+			<artifactId>zeebe-client-java</artifactId>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.testcontainers</groupId>
+			<artifactId>junit-jupiter</artifactId>
+			<scope>test</scope>
+		</dependency>
+
+		<dependency>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter</artifactId>
 			<scope>test</scope>
 		</dependency>
 
@@ -42,7 +73,7 @@
 		<dependency>
 			<groupId>org.apache.logging.log4j</groupId>
 			<artifactId>log4j-slf4j-impl</artifactId>
-			<version>2.11.1</version>
+			<version>2.17.1</version>
 		</dependency>
 
 	</dependencies>
@@ -78,9 +109,9 @@
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
+				<version>3.9.0</version>
 				<configuration>
-					<source>1.8</source>
-					<target>1.8</target>
+					<release>17</release>
 				</configuration>
 			</plugin>
 		</plugins>
